Krishibodh-Robot-Core/camera/requirements.txt
opencv-python
rembg
pillow
onnxruntime
